Update requirements to last version
`@gen_test`, the now recommended way to write asynchronous test
automatically times out after a delay of 5 seconds, so there is no need
for the `assert_eventually_equals` function in this case.

The `yield` keyword does not mean the same thing when used inside a
regular function or inside a function tagged with `gen_test` (in which
case it is more or less equivalent to the `await` keyword) so I was
forced to "defactor" the `_get_delivery_status` and `_get_size`
functions.