Made configs compatible with pythons yaml handling
Apparently the curly brackets in the configuration files are not handled
as strings by the python yaml-library that I use for reading the
configs. Values needs to be explicitly specified as strings by using "<string>" otherwise the code will crash when it finds a curly bracket. However, regular expressions can not be within "<reg-exp>".